add backend for US05 and US06
DELETE to /tables/:table_id/seat  updates by removing the
reservation_id for the requested table_id. And also changes reservation
status to "finished."
POST /reservations  changed to return 201 if status is "booked" (which
is the default status), and return 400 if status is "seated" or
"finished".
PUT /reservations/:reservation_id/status  now returns 200 for approved
statuses, or returns 400 for status of "unknown".
Some knex transaction methods added to sync updates to "tables" and
"reservations" tables with each request.